Se actualizan las interfaces de repositorio para cambiar el tipo de dato de los identificadores de Long a Integer, mejorando la consistencia en la gestión de identificadores.

+++ back_local/src/main/java/com/portfolioweb/aj/Repositorio/RHabilidad.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RHabilidad extends JpaRepository<Habilidad, Long> {
+public interface RHabilidad extends JpaRepository<Habilidad, Integer> {
 
     Optional<Habilidad> findByNombre(String nombre);
