+++ back_local/src/main/java/com/portfolioweb/aj/Repositorio/RTipoEducacion.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RTipoEducacion extends JpaRepository<TipoEducacion, Long> {
+public interface RTipoEducacion extends JpaRepository<TipoEducacion, Integer> {
 }
